requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers

